some naming fix
