+++ analytical_engine/java/grape-jdk/src/main/java/com/alibaba/graphscope/communication/FFICommunicator.java
@@ -30,8 +30,8 @@ import com.alibaba.fastffi.FFIPointer;
 import com.alibaba.fastffi.FFITypeAlias;
 
 /**
- * The java wrapper for <a
- * href="https://github.com/alibaba/libgrape-lite/blob/master/grape/communication/communicator.h">grape
+ * The java wrapper for <a href=
+ * "https://github.com/alibaba/libgrape-lite/blob/master/grape/communication/communicator.h">grape
  * Communicator class</a>, which provides some useful aggregators. Shall not be used by user.
  *
  * <p>For User-defined msg types, the corrsponding code is generated by graphscope-processor.
@@ -41,18 +41,24 @@ import com.alibaba.fastffi.FFITypeAlias;
 @CXXHead(CORE_JAVA_JAVA_MESSAGES_H)
 @FFITypeAlias(GRAPE_COMMUNICATOR)
 public interface FFICommunicator extends FFIPointer {
-    // @CXXTemplate(cxx = DOUBLE_MSG, java = "com.alibaba.grape.parallel.message.DoubleMsg")
-    // @CXXTemplate(cxx = LONG_MSG, java = "com.alibaba.grape.parallel.message.LongMsg")
+    // @CXXTemplate(cxx = DOUBLE_MSG, java =
+    // "com.alibaba.graphscope.parallel.message.DoubleMsg")
+    // @CXXTemplate(cxx = LONG_MSG, java =
+    // "com.alibaba.graphscope.parallel.message.LongMsg")
     @FFINameAlias("Sum")
     <MSG_T> void sum(@FFIConst @CXXReference MSG_T msgIn, @CXXReference MSG_T msgOut);
 
-    // @CXXTemplate(cxx = DOUBLE_MSG, java = "com.alibaba.grape.parallel.message.DoubleMsg")
-    // @CXXTemplate(cxx = LONG_MSG, java = "com.alibaba.grape.parallel.message.LongMsg")
+    // @CXXTemplate(cxx = DOUBLE_MSG, java =
+    // "com.alibaba.graphscope.parallel.message.DoubleMsg")
+    // @CXXTemplate(cxx = LONG_MSG, java =
+    // "com.alibaba.graphscope.parallel.message.LongMsg")
     @FFINameAlias("Min")
     <MSG_T> void min(@FFIConst @CXXReference MSG_T msgIn, @CXXReference MSG_T msgOut);
 
-    // @CXXTemplate(cxx = DOUBLE_MSG, java = "com.alibaba.grape.parallel.message.DoubleMsg")
-    // @CXXTemplate(cxx = LONG_MSG, java = "com.alibaba.grape.parallel.message.LongMsg")
+    // @CXXTemplate(cxx = DOUBLE_MSG, java =
+    // "com.alibaba.graphscope.parallel.message.DoubleMsg")
+    // @CXXTemplate(cxx = LONG_MSG, java =
+    // "com.alibaba.graphscope.parallel.message.LongMsg")
     @FFINameAlias("Max")
     <MSG_T> void max(@FFIConst @CXXReference MSG_T msgIn, @CXXReference MSG_T msgOut);
 }
