+++ analytical_engine/java/grape-jdk/src/main/java/com/alibaba/graphscope/context/ffi/FFIVertexDataContext.java
@@ -34,27 +34,30 @@ import com.alibaba.graphscope.utils.CppHeaderName;
 @CXXHead(ARROW_PROJECTED_FRAGMENT_H)
 @FFITypeAlias(VERTEX_DATA_CONTEXT)
 // @CXXTemplate(
-//        cxx = {
-//            ARROW_PROJECTED_FRAGMENT + "<int64_t,uint64_t," + GRAPE_EMPTY_TYPE + ",int64_t>",
-//            "double"
-//        },
-//        java = {
+// cxx = {
+// ARROW_PROJECTED_FRAGMENT + "<int64_t,uint64_t," + GRAPE_EMPTY_TYPE +
+// ",int64_t>",
+// "double"
+// },
+// java = {
 //
-// "com.alibaba.graphscope.fragment.ArrowProjectedFragment<Long,Long,com.alibaba.grape.ds.EmptyType,Long>",
-//            "java.lang.Double"
-//        })
+// "com.alibaba.graphscope.fragment.ArrowProjectedFragment<Long,Long,com.alibaba.graphscope.ds.EmptyType,Long>",
+// "java.lang.Double"
+// })
 // @CXXTemplate(
-//        cxx = {ARROW_PROJECTED_FRAGMENT + "<int64_t,uint64_t,double,int64_t>", "double"},
-//        java = {
-//            "com.alibaba.graphscope.fragment.ArrowProjectedFragment<Long,Long,Double,Long>",
-//            "java.lang.Double"
-//        })
+// cxx = {ARROW_PROJECTED_FRAGMENT + "<int64_t,uint64_t,double,int64_t>",
+// "double"},
+// java = {
+// "com.alibaba.graphscope.fragment.ArrowProjectedFragment<Long,Long,Double,Long>",
+// "java.lang.Double"
+// })
 // @CXXTemplate(
-//        cxx = {ARROW_PROJECTED_FRAGMENT + "<int64_t,uint64_t,double,int64_t>", "int64_t"},
-//        java = {
-//            "com.alibaba.graphscope.fragment.ArrowProjectedFragment<Long,Long,Double,Long>",
-//            "java.lang.Long"
-//        })
+// cxx = {ARROW_PROJECTED_FRAGMENT + "<int64_t,uint64_t,double,int64_t>",
+// "int64_t"},
+// java = {
+// "com.alibaba.graphscope.fragment.ArrowProjectedFragment<Long,Long,Double,Long>",
+// "java.lang.Long"
+// })
 public interface FFIVertexDataContext<FRAG_T, DATA_T> extends FFIPointer {
 
     @CXXReference
