+++ analytical_engine/java/grape-jdk/src/main/java/com/alibaba/graphscope/parallel/PropertyMessageManager.java
@@ -40,18 +40,6 @@ import com.alibaba.graphscope.utils.JNILibraryName;
 @FFITypeAlias(CppClassName.PROPERTY_MESSAGE_MANAGER)
 public interface PropertyMessageManager extends MessageManagerBase {
 
-    //     @CXXTemplate(
-    //                cxx = {CppClassName.ARROW_FRAGMENT + "<int64_t>", DOUBLE_MSG},
-    //                java = {
-    //                    "com.alibaba.graphscope.fragment.ArrowFragment<java.lang.Long>",
-    //                    "com.alibaba.grape.parallel.message.DoubleMsg"
-    //                })
-    //     @CXXTemplate(
-    //                cxx = {CppClassName.ARROW_FRAGMENT + "<int64_t>", LONG_MSG},
-    //                java = {
-    //                    "com.alibaba.graphscope.fragment.ArrowFragment<java.lang.Long>",
-    //                    "com.alibaba.grape.parallel.message.LongMsg"
-    //                })
     @FFINameAlias("SendMsgThroughIEdges")
     <FRAG_T, MSG_T> void sendMsgThroughIEdges(
             @CXXReference FRAG_T frag,
@@ -59,18 +47,6 @@ public interface PropertyMessageManager extends MessageManagerBase {
             int eLabelId,
             @CXXReference MSG_T msg);
 
-    // @CXXTemplate(
-    //            cxx = {CppClassName.ARROW_FRAGMENT + "<int64_t>", DOUBLE_MSG},
-    //            java = {
-    //                "com.alibaba.graphscope.fragment.ArrowFragment<java.lang.Long>",
-    //                "com.alibaba.grape.parallel.message.DoubleMsg"
-    //            })
-    // @CXXTemplate(
-    //            cxx = {CppClassName.ARROW_FRAGMENT + "<int64_t>", LONG_MSG},
-    //            java = {
-    //                "com.alibaba.graphscope.fragment.ArrowFragment<java.lang.Long>",
-    //                "com.alibaba.grape.parallel.message.LongMsg"
-    //            })
     @FFINameAlias("SendMsgThroughOEdges")
     <FRAG_T, MSG_T> void sendMsgThroughOEdges(
             @CXXReference FRAG_T frag,
@@ -78,18 +54,6 @@ public interface PropertyMessageManager extends MessageManagerBase {
             int eLabelId,
             @CXXReference MSG_T msg);
 
-    // @CXXTemplate(
-    //            cxx = {CppClassName.ARROW_FRAGMENT + "<int64_t>", DOUBLE_MSG},
-    //            java = {
-    //                "com.alibaba.graphscope.fragment.ArrowFragment<java.lang.Long>",
-    //                "com.alibaba.grape.parallel.message.DoubleMsg"
-    //            })
-    // @CXXTemplate(
-    //            cxx = {CppClassName.ARROW_FRAGMENT + "<int64_t>", LONG_MSG},
-    //            java = {
-    //                "com.alibaba.graphscope.fragment.ArrowFragment<java.lang.Long>",
-    //                "com.alibaba.grape.parallel.message.LongMsg"
-    //            })
     @FFINameAlias("SendMsgThroughEdges")
     <FRAG_T, MSG_T> void sendMsgThroughEdges(
             @CXXReference FRAG_T frag,
@@ -97,36 +61,12 @@ public interface PropertyMessageManager extends MessageManagerBase {
             int eLabelId,
             @CXXReference MSG_T msg);
 
-    // @CXXTemplate(
-    //            cxx = {CppClassName.ARROW_FRAGMENT + "<int64_t>", DOUBLE_MSG},
-    //            java = {
-    //                "com.alibaba.graphscope.fragment.ArrowFragment<java.lang.Long>",
-    //                "com.alibaba.grape.parallel.message.DoubleMsg"
-    //            })
-    // @CXXTemplate(
-    //            cxx = {CppClassName.ARROW_FRAGMENT + "<int64_t>", LONG_MSG},
-    //            java = {
-    //                "com.alibaba.graphscope.fragment.ArrowFragment<java.lang.Long>",
-    //                "com.alibaba.grape.parallel.message.LongMsg"
-    //            })
     @FFINameAlias("SyncStateOnOuterVertex")
     <FRAG_T, MSG_T> void syncStateOnOuterVertex(
             @CXXReference FRAG_T frag,
             @CXXReference @FFITypeAlias(GRAPE_LONG_VERTEX) Vertex<Long> vertex,
             @CXXReference MSG_T msg);
 
-    //    @CXXTemplate(
-    //            cxx = {CppClassName.ARROW_FRAGMENT + "<int64_t>", DOUBLE_MSG},
-    //            java = {
-    //                "com.alibaba.graphscope.fragment.ArrowFragment<java.lang.Long>",
-    //                "com.alibaba.graphscope.parallel.message.DoubleMsg"
-    //            })
-    //    @CXXTemplate(
-    //            cxx = {CppClassName.ARROW_FRAGMENT + "<int64_t>", LONG_MSG},
-    //            java = {
-    //                "com.alibaba.graphscope.fragment.ArrowFragment<java.lang.Long>",
-    //                "com.alibaba.graphscope.parallel.message.LongMsg"
-    //            })
     @FFINameAlias("GetMessage")
     <FRAG_T, MSG_T> boolean getMessage(
             @CXXReference FRAG_T frag,
