+++ analytical_engine/java/grape-runtime/src/main/java/com/alibaba/graphscope/annotation/GraphScopeAppScanner.java
@@ -122,15 +122,16 @@ public class GraphScopeAppScanner {
 
     private String scanAppAndGenerateImpl() {
         // readConfig();
-        // for all ffiMirror we suppose they are all possible to appear in messageManager
+        // for all ffiMirror we suppose they are all possible to appear in
+        // messageManager
         collectFFIMirror();
         StringBuilder sb = new StringBuilder();
         for (String str : ffiMirrors.keySet()) {
             sb.append(ffiMirrors.get(str) + "=" + str + ",");
         }
         // add support for doubleMsg and LongMsg
-        //        sb.append("gs::DoubleMsg=com.alibaba.grapescope.parallel.message.DoubleMsg,");
-        //        sb.append("gs::LongMsg=com.alibaba.grapescope.parallel.message.LongMsg,");
+        // sb.append("gs::DoubleMsg=com.alibaba.graphscopescope.parallel.message.DoubleMsg,");
+        // sb.append("gs::LongMsg=com.alibaba.graphscopescope.parallel.message.LongMsg,");
         String temp = sb.toString();
         String messageTypes = temp.substring(0, temp.length() - 1);
         // create a dummy graphConfig, oid type should never be used.
@@ -374,7 +375,7 @@ public class GraphScopeAppScanner {
                     String className = relative.toString();
                     className = className.substring(0, className.length() - 6).replace('/', '.');
                     // For FFIMirror found in grape-sdk, we don't count it in codegeneration
-                    if (className.contains("com.alibaba.grape.stdcxx.FFISample")) {
+                    if (className.contains("com.alibaba.graphscope.stdcxx.FFISample")) {
                         logger.info("catch ffi mirror FFISample, skip generation" + className);
                         return FileVisitResult.CONTINUE;
                     }
