+++ analytical_engine/java/grape-runtime/src/test/java/com/alibaba/graphscope/ds/DenseVertexSetTest.java
@@ -14,11 +14,8 @@
  * limitations under the License.
  */
 
-package com.alibaba.grape.ds;
+package com.alibaba.graphscope.ds;
 
-import com.alibaba.graphscope.ds.DenseVertexSet;
-import com.alibaba.graphscope.ds.Vertex;
-import com.alibaba.graphscope.ds.VertexRange;
 import com.alibaba.graphscope.utils.FFITypeFactoryhelper;
 import org.junit.Assert;
 import org.junit.Test;
