+++ analytical_engine/java/grape-runtime/src/test/java/com/alibaba/graphscope/stdcxx/StdUnorderedMapTest.java
@@ -14,33 +14,34 @@
  * limitations under the License.
  */
 
-package com.alibaba.grape.ds;
+package com.alibaba.graphscope.stdcxx;
 
-import com.alibaba.graphscope.ds.Bitset;
+import com.alibaba.fastffi.FFITypeFactory;
+import com.alibaba.graphscope.stdcxx.StdUnorderedMap.Factory;
+import com.alibaba.graphscope.utils.CppClassName;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
-public class BitsetTest {
+public class StdUnorderedMapTest {
 
-    private Bitset bitset;
+    private Factory<Integer, Long> factory =
+            FFITypeFactory.getFactory(
+                    StdUnorderedMap.class, CppClassName.STD_UNORDERED_MAP + "<unsigned,uint64_t>");
+    private StdUnorderedMap<Integer, Long> map;
 
     @Before
-    public void before() {
-        bitset = Bitset.factory.create();
-        bitset.init(20);
+    public void init() {
+        map = factory.create();
     }
 
     @Test
     public void test1() {
-        for (int i = 0; i < 10; i += 2) {
-            bitset.setBit((long) i);
+        for (int i = 0; i < 10; ++i) {
+            map.set(i, (long) i);
         }
-        for (int i = 0; i < 10; i += 2) {
-            Assert.assertTrue(bitset.getBit((long) i));
-        }
-        for (int i = 1; i < 10; i += 2) {
-            Assert.assertFalse(bitset.getBit((long) i));
+        for (int i = 0; i < 10; ++i) {
+            Assert.assertTrue("Different at " + i, i == map.get(i).intValue());
         }
     }
 }
